Performance changes. attempting to mitigate memory leaks from the renderer.
After lots of research and log tests I determined the performance issues are due to a memory leak in the OpenGlRenderer or some other component in the renderer pipeline. It looks like my code running in the background executes in under 15 ms thus it isn't a problem with the code, but the memory blueprint for rendering the screen keeps increasing with every press of the button. Until more work is done on the .net maui platform I think we will just have to deal with the problem by restarting the app after a few generations of the name gen or once it becomes too chuggy.